initial commit
This code went through a couple of refactorings, but I managed to destroy the git history...

the first iteration used context and hooks to manage state.
then it was refactored to use redux. that was a bit messy, so I reorganized the code and brought in immer to help make my reducers more readable